Refatorando validação de erros. -> Criando uma função para validar todos os erros  nos valores passados por parametros.
